# Improve language detection performance