--- a/pkg/distributor/model/push.go
+++ b/pkg/distributor/model/push.go
@@ -2,6 +2,7 @@ package model
 
 import (
 	v1 "github.com/grafana/pyroscope/api/gen/proto/go/types/v1"
+	phlaremodel "github.com/grafana/pyroscope/pkg/model"
 	"github.com/grafana/pyroscope/pkg/pprof"
 )
 
@@ -29,4 +30,38 @@ type ProfileSample struct {
 type ProfileSeries struct {
 	Labels  []*v1.LabelPair
 	Samples []*ProfileSample
+}
+
+func (p *ProfileSeries) GetLanguage() string {
+	spyName := phlaremodel.Labels(p.Labels).Get(phlaremodel.LabelNamePyroscopeSpy)
+	if spyName != "" {
+		lang := getProfileLanguageFromSpy(spyName)
+		if lang != "" {
+			return lang
+		}
+	}
+	return ""
+}
+
+func getProfileLanguageFromSpy(spyName string) string {
+	switch spyName {
+	default:
+		return ""
+	case "dotnetspy":
+		return "dotnet"
+	case "gospy":
+		return "go"
+	case "phpspy":
+		return "php"
+	case "pyspy":
+		return "python"
+	case "rbspy":
+		return "ruby"
+	case "nodespy":
+		return "nodejs"
+	case "javaspy":
+		return "java"
+	case "pyroscope-rs":
+		return "rust"
+	}
 }