--- a/pkg/ingester/pyroscope/ingest_adapter.go
+++ b/pkg/ingester/pyroscope/ingest_adapter.go
@@ -118,7 +118,7 @@ func (p *pyroscopeIngesterAdapter) Put(ctx context.Context, pi *storage.PutInput
 	})
 	if pi.SpyName != "" {
 		series.Labels = append(series.Labels, &typesv1.LabelPair{
-			Name:  "pyroscope_spy",
+			Name:  phlaremodel.LabelNamePyroscopeSpy,
 			Value: pi.SpyName,
 		})
 	}