--- a/pkg/model/labels.go
+++ b/pkg/model/labels.go
@@ -22,15 +22,16 @@ import (
 var seps = []byte{'\xff'}
 
 const (
-	LabelNameProfileType = "__profile_type__"
-	LabelNameType        = "__type__"
-	LabelNameUnit        = "__unit__"
-	LabelNamePeriodType  = "__period_type__"
-	LabelNamePeriodUnit  = "__period_unit__"
-	LabelNameDelta       = "__delta__"
-	LabelNameProfileName = pmodel.MetricNameLabel
-	LabelNameServiceName = "service_name"
-	LabelNameSessionID   = "__session_id__"
+	LabelNameProfileType  = "__profile_type__"
+	LabelNameType         = "__type__"
+	LabelNameUnit         = "__unit__"
+	LabelNamePeriodType   = "__period_type__"
+	LabelNamePeriodUnit   = "__period_unit__"
+	LabelNameDelta        = "__delta__"
+	LabelNameProfileName  = pmodel.MetricNameLabel
+	LabelNameServiceName  = "service_name"
+	LabelNamePyroscopeSpy = "pyroscope_spy"
+	LabelNameSessionID    = "__session_id__"
 
 	LabelNameServiceNameK8s = "__meta_kubernetes_pod_annotation_pyroscope_io_service_name"