--- a/pkg/og/convert/pprof/profile.go
+++ b/pkg/og/convert/pprof/profile.go
@@ -211,7 +211,7 @@ func (p *RawProfile) createLabels(profile *pprof.Profile, md ingestion.Metadata)
 		Name:  "service_name",
 		Value: md.Key.AppName(),
 	}, &v1.LabelPair{
-		Name:  "pyroscope_spy",
+		Name:  phlaremodel.LabelNamePyroscopeSpy,
 		Value: md.SpyName,
 	})
 	for k, v := range md.Key.Labels() {