@korniltsev 2023-12-16
There may or may be not a new spy label from grafana agent pyroscope.java component

https://github.com/grafana/agent/pull/5985/files#diff-a0c3d4195b76376947e1f9e0059dc3538e97c3fb10ba70d9000ea74b902c1f8dR244

Let me know if you have any preference for its value or its presence