@aleks-p 2023-12-18
Thanks @korniltsev, the value makes sense to me. I'll add the mapping for it.